+++ darpa_proposal_generator/__init__.py
@@ -1,0 +1,1 @@
+# This file makes the darpa_proposal_generator directory a Python package.
