+++ darpa_proposal_generator/roadmap.py
@@ -1,0 +1,301 @@
+from typing import Dict
+
+class ExperimentalValidationRoadmap:
+    """
+    Comprehensive experimental validation plan with realistic milestones
+    """
+
+    def __init__(self):
+        self.current_trl = 2
+        self.validation_phases = self.define_validation_phases()
+
+    def define_validation_phases(self) -> Dict:
+        """Define realistic experimental validation roadmap"""
+
+        phases = {
+            "phase_1_theoretical_validation": {
+                "duration_months": 6,
+                "trl_start": 2,
+                "trl_end": 3,
+                "budget": 1.5e6,
+                "key_experiments": [
+                    {
+                        "experiment": "Harmonic oscillator state preparation",
+                        "platform": "Superconducting qubits (IBM Quantum)",
+                        "success_metric": "99% state preparation fidelity",
+                        "timeline_months": 3,
+                        "risk_level": "Low"
+                    },
+                    {
+                        "experiment": "Vibrational state superposition",
+                        "platform": "Trapped ions (IonQ access)",
+                        "success_metric": "Coherent superposition of 5+ vibrational modes",
+                        "timeline_months": 4,
+                        "risk_level": "Medium"
+                    },
+                    {
+                        "experiment": "Position encoding demonstration",
+                        "platform": "Quantum simulator (classical)",
+                        "success_metric": "Theoretical 10cm localization accuracy",
+                        "timeline_months": 2,
+                        "risk_level": "Low"
+                    }
+                ],
+                "deliverables": [
+                    "Peer-reviewed publication on vibrational localization theory",
+                    "Quantum circuit implementations for position encoding",
+                    "Simulation framework validation"
+                ],
+                "success_criteria": "All key experiments achieve >80% of target metrics"
+            },
+
+            "phase_2_laboratory_demonstration": {
+                "duration_months": 12,
+                "trl_start": 3,
+                "trl_end": 4,
+                "budget": 8e6,
+                "key_experiments": [
+                    {
+                        "experiment": "Quantum localization in controlled environment",
+                        "platform": "Custom superconducting qubit system",
+                        "success_metric": "1m localization accuracy in lab",
+                        "timeline_months": 8,
+                        "risk_level": "High"
+                    },
+                    {
+                        "experiment": "Decoherence characterization",
+                        "platform": "Multiple quantum platforms",
+                        "success_metric": "100μs coherence time for localization",
+                        "timeline_months": 6,
+                        "risk_level": "Medium"
+                    },
+                    {
+                        "experiment": "Classical-quantum interface",
+                        "platform": "Hybrid classical-quantum system",
+                        "success_metric": "Real-time position updates at 1Hz",
+                        "timeline_months": 10,
+                        "risk_level": "Medium"
+                    },
+                    {
+                        "experiment": "Noise resilience testing",
+                        "platform": "Laboratory with controlled noise",
+                        "success_metric": "Maintains accuracy under 30dB interference",
+                        "timeline_months": 9,
+                        "risk_level": "Medium"
+                    }
+                ],
+                "deliverables": [
+                    "Laboratory prototype quantum localization system",
+                    "Performance characterization report",
+                    "Integration interface specifications"
+                ],
+                "success_criteria": "Demonstrate quantum advantage in controlled laboratory environment"
+            },
+
+            "phase_3_relevant_environment": {
+                "duration_months": 18,
+                "trl_start": 4,
+                "trl_end": 5,
+                "budget": 15e6,
+                "key_experiments": [
+                    {
+                        "experiment": "Environmental robustness testing",
+                        "platform": "Ruggedized prototype",
+                        "success_metric": "Operation in -40°C to +70°C range",
+                        "timeline_months": 12,
+                        "risk_level": "High"
+                    },
+                    {
+                        "experiment": "Mobile platform integration",
+                        "platform": "Vehicle-mounted system",
+                        "success_metric": "3m accuracy during vehicle motion",
+                        "timeline_months": 15,
+                        "risk_level": "High"
+                    },
+                    {
+                        "experiment": "GPS-denied environment testing",
+                        "platform": "Underground/shielded facility",
+                        "success_metric": "Maintain localization without GPS",
+                        "timeline_months": 14,
+                        "risk_level": "Medium"
+                    },
+                    {
+                        "experiment": "Electromagnetic interference testing",
+                        "platform": "Anechoic chamber with RF sources",
+                        "success_metric": "Immune to 60dB jamming signals",
+                        "timeline_months": 10,
+                        "risk_level": "Medium"
+                    }
+                ],
+                "deliverables": [
+                    "Ruggedized quantum localization prototype",
+                    "Military environment test report",
+                    "Integration guidelines for defense platforms"
+                ],
+                "success_criteria": "Demonstrate reliable operation in relevant military environments"
+            },
+
+            "phase_4_operational_environment": {
+                "duration_months": 24,
+                "trl_start": 5,
+                "trl_end": 6,
+                "budget": 20e6,
+                "key_experiments": [
+                    {
+                        "experiment": "Submarine navigation trial",
+                        "platform": "Navy submarine (if approved)",
+                        "success_metric": "10m accuracy over 24-hour dive",
+                        "timeline_months": 20,
+                        "risk_level": "Very High"
+                    },
+                    {
+                        "experiment": "Aircraft navigation trial",
+                        "platform": "Military aircraft testbed",
+                        "success_metric": "5m accuracy during flight operations",
+                        "timeline_months": 18,
+                        "risk_level": "Very High"
+                    },
+                    {
+                        "experiment": "Ground vehicle navigation trial",
+                        "platform": "Army ground vehicle",
+                        "success_metric": "3m accuracy in urban environment",
+                        "timeline_months": 15,
+                        "risk_level": "High"
+                    }
+                ],
+                "deliverables": [
+                    "Operational prototype systems",
+                    "Military field test results",
+                    "Technology transfer package"
+                ],
+                "success_criteria": "Successful operational demonstrations with military partners"
+            }
+        }
+
+        return phases
+
+    def generate_risk_mitigation_plan(self) -> Dict:
+        """Generate comprehensive risk mitigation strategies"""
+
+        risk_mitigation = {
+            "technical_risks": {
+                "decoherence_limitations": {
+                    "risk_description": "Quantum coherence times insufficient for practical navigation",
+                    "probability": "Medium (40%)",
+                    "impact": "High (delays deployment by 12-24 months)",
+                    "mitigation_strategies": [
+                        "Parallel development on multiple quantum platforms",
+                        "Conservative performance targets with margin",
+                        "Hybrid classical-quantum algorithms",
+                        "Error correction protocol development"
+                    ],
+                    "contingency_plan": "Fall back to quantum-enhanced classical navigation",
+                    "early_warning_indicators": [
+                        "Coherence times <50μs in lab conditions",
+                        "Decoherence rate increases >10x in mobile platform"
+                    ]
+                },
+                "integration_complexity": {
+                    "risk_description": "Quantum system integration with classical navigation proves too complex",
+                    "probability": "High (60%)",
+                    "impact": "Medium (increases cost by 25-50%)",
+                    "mitigation_strategies": [
+                        "Modular system architecture",
+                        "Standard interface development",
+                        "Incremental integration approach",
+                        "Classical backup systems"
+                    ],
+                    "contingency_plan": "Standalone quantum navigation system",
+                    "early_warning_indicators": [
+                        "Interface development >6 months behind schedule",
+                        "Power consumption >5x classical systems"
+                    ]
+                },
+                "manufacturing_scalability": {
+                    "risk_description": "Quantum hardware cannot be manufactured at scale",
+                    "probability": "Medium (30%)",
+                    "impact": "High (limits market penetration)",
+                    "mitigation_strategies": [
+                        "Partnership with quantum hardware manufacturers",
+                        "Standardized component selection",
+                        "Volume manufacturing agreements",
+                        "Alternative fabrication approaches"
+                    ],
+                    "contingency_plan": "Limited production for niche applications",
+                    "early_warning_indicators": [
+                        "Unit costs >$1M per system",
+                        "Manufacturing yield <50%"
+                    ]
+                }
+            },
+            "programmatic_risks": {
+                "funding_continuity": {
+                    "risk_description": "Government funding reduced or cancelled",
+                    "probability": "Low (20%)",
+                    "impact": "Very High (program termination)",
+                    "mitigation_strategies": [
+                        "Multiple funding sources (DARPA, ONR, AFRL, ARL)",
+                        "Private investor engagement",
+                        "Commercial application development",
+                        "International partnership opportunities"
+                    ],
+                    "contingency_plan": "Transition to commercial-focused development",
+                    "early_warning_indicators": [
+                        "Budget cuts >25% in any fiscal year",
+                        "Program manager changes >2 times"
+                    ]
+                },
+                "team_retention": {
+                    "risk_description": "Key technical personnel leave for other opportunities",
+                    "probability": "Medium (35%)",
+                    "impact": "Medium (delays and knowledge loss)",
+                    "mitigation_strategies": [
+                        "Competitive compensation packages",
+                        "Equity participation for key personnel",
+                        "Professional development opportunities",
+                        "Flexible work arrangements"
+                    ],
+                    "contingency_plan": "Rapid hiring and knowledge transfer protocols",
+                    "early_warning_indicators": [
+                        "Turnover rate >20% annually",
+                        "Key personnel expressing dissatisfaction"
+                    ]
+                },
+                "competition_advancement": {
+                    "risk_description": "Competitors achieve similar capabilities first",
+                    "probability": "Medium (40%)",
+                    "impact": "High (reduced market opportunity)",
+                    "mitigation_strategies": [
+                        "Accelerated development timeline",
+                        "Patent protection strategy",
+                        "First-mover advantage in military applications",
+                        "Unique differentiating features"
+                    ],
+                    "contingency_plan": "Focus on superior performance or cost advantages",
+                    "early_warning_indicators": [
+                        "Competitor quantum navigation announcements",
+                        "Similar DARPA awards to other teams"
+                    ]
+                }
+            },
+            "market_risks": {
+                "military_adoption_rate": {
+                    "risk_description": "Military adoption slower than projected",
+                    "probability": "High (70%)",
+                    "impact": "Medium (reduced revenue timeline)",
+                    "mitigation_strategies": [
+                        "Conservative adoption projections",
+                        "Early military partner engagement",
+                        "Demonstration of clear operational advantages",
+                        "Integration with existing procurement programs"
+                    ],
+                    "contingency_plan": "Focus on civilian applications first",
+                    "early_warning_indicators": [
+                        "Military feedback indicates low interest",
+                        "Procurement timeline >2 years longer than expected"
+                    ]
+                }
+            }
+        }
+
+        return risk_mitigation
