+++ darpa_proposal_generator/theory.py
@@ -1,0 +1,165 @@
+import numpy as np
+from scipy.special import hermite
+from qiskit_aer.noise import NoiseModel, depolarizing_error, amplitude_damping_error, phase_damping_error
+import logging
+from typing import Dict, List, Tuple, Optional
+
+logger = logging.getLogger(__name__)
+
+class QuantumLocalizationTheory:
+    """
+    Rigorous theoretical foundation for vibrational state localization
+
+    Based on the mathematical framework where spatial coordinates are encoded
+    in the quantum harmonic oscillator basis states, allowing position
+    information to be manipulated through vibrational quantum numbers.
+    """
+
+    def __init__(self, omega: float = 1.0, hbar: float = 1.0, mass: float = 1.0):
+        """
+        Initialize theoretical framework parameters
+
+        Args:
+            omega: Harmonic oscillator frequency
+            hbar: Reduced Planck constant (natural units)
+            mass: Effective mass parameter
+        """
+        self.omega = omega
+        self.hbar = hbar
+        self.mass = mass
+        self.x0 = np.sqrt(hbar / (mass * omega))  # Characteristic length scale
+
+    def harmonic_oscillator_wavefunction(self, n: int, x: np.ndarray) -> np.ndarray:
+        """
+        Generate quantum harmonic oscillator wavefunction for state |n⟩
+
+        ψₙ(x) = (mω/πℏ)^(1/4) * (1/√(2ⁿn!)) * Hₙ(x/x₀) * exp(-x²/(2x₀²))
+        """
+        # Normalization constant
+        norm = (self.mass * self.omega / (np.pi * self.hbar))**(1/4)
+        norm *= 1.0 / np.sqrt(2**n * np.math.factorial(n))
+
+        # Dimensionless coordinate
+        xi = x / self.x0
+
+        # Hermite polynomial
+        hermite_poly = hermite(n)
+        Hn = hermite_poly(xi)
+
+        # Gaussian envelope
+        gaussian = np.exp(-xi**2 / 2)
+
+        return norm * Hn * gaussian
+
+    def vibrational_coordinate_encoding(self, position: float, max_n: int = 10) -> np.ndarray:
+        """
+        Encode spatial position as superposition of vibrational states
+
+        |ψ(x₀)⟩ = Σₙ cₙ|n⟩ where cₙ are chosen to localize at position x₀
+
+        This is the core theoretical innovation: position becomes a quantum
+        observable through the vibrational quantum number basis.
+        """
+        # Calculate coefficients to maximize localization at target position
+        coefficients = np.zeros(max_n + 1, dtype=complex)
+
+        # Use coherent state approach: |α⟩ = e^(-|α|²/2) Σₙ (αⁿ/√n!)|n⟩
+        alpha = position / self.x0  # Dimensionless displacement parameter
+
+        for n in range(max_n + 1):
+            coefficients[n] = (alpha**n / np.sqrt(np.math.factorial(n))) * np.exp(-abs(alpha)**2 / 2)
+
+        # Normalize
+        norm = np.sqrt(np.sum(np.abs(coefficients)**2))
+        coefficients /= norm
+
+        return coefficients
+
+    def theoretical_position_uncertainty(self, coefficients: np.ndarray) -> float:
+        """
+        Calculate theoretical minimum position uncertainty for given state
+
+        Δx = x₀ * √⟨n̂⟩ for harmonic oscillator states
+        """
+        n_values = np.arange(len(coefficients))
+        mean_n = np.sum(np.abs(coefficients)**2 * n_values)
+        return self.x0 * np.sqrt(mean_n + 0.5)  # Include zero-point motion
+
+class RealisticQuantumSystem:
+    """
+    Model realistic quantum system with proper decoherence and error sources
+    """
+
+    def __init__(self, platform_type: str = "superconducting"):
+        """
+        Initialize with realistic parameters for different quantum platforms
+        """
+        self.platform_type = platform_type
+        self.setup_platform_parameters()
+
+    def setup_platform_parameters(self):
+        """Set realistic parameters based on current quantum technology"""
+
+        if self.platform_type == "superconducting":
+            # IBM/Google superconducting qubit parameters (2024 state-of-art)
+            self.T1 = 100e-6  # Energy relaxation time (100 μs)
+            self.T2 = 50e-6   # Dephasing time (50 μs)
+            self.gate_time_1q = 30e-9  # Single-qubit gate time (30 ns)
+            self.gate_time_2q = 200e-9  # Two-qubit gate time (200 ns)
+            self.gate_error_1q = 1e-4   # Single-qubit gate error
+            self.gate_error_2q = 5e-3   # Two-qubit gate error
+            self.readout_error = 2e-2   # Measurement error
+            self.operating_temp = 0.015  # 15 mK
+
+        elif self.platform_type == "trapped_ion":
+            # IonQ/Honeywell trapped ion parameters
+            self.T1 = 10.0    # Very long coherence (10 s)
+            self.T2 = 1.0     # Dephasing time (1 s)
+            self.gate_time_1q = 10e-6   # Single-qubit gate time (10 μs)
+            self.gate_time_2q = 100e-6  # Two-qubit gate time (100 μs)
+            self.gate_error_1q = 1e-5   # Excellent single-qubit fidelity
+            self.gate_error_2q = 1e-3   # Good two-qubit fidelity
+            self.readout_error = 1e-3   # Excellent readout
+            self.operating_temp = 1e-6  # μK effective temperature
+
+        elif self.platform_type == "photonic":
+            # Xanadu/PsiQuantum photonic parameters
+            self.T1 = np.inf  # No energy relaxation for photons
+            self.T2 = 1e-3    # Limited by detection efficiency
+            self.gate_time_1q = 1e-12   # Speed of light limited (1 ps)
+            self.gate_time_2q = 1e-9    # Limited by nonlinear optics (1 ns)
+            self.gate_error_1q = 1e-3   # Limited by imperfect components
+            self.gate_error_2q = 1e-1   # Challenging two-qubit gates
+            self.readout_error = 1e-1   # Detector efficiency ~90%
+            self.operating_temp = 300   # Room temperature operation
+
+        logger.info(f"Initialized {self.platform_type} quantum system with T1={self.T1:.2e}s, T2={self.T2:.2e}s")
+
+    def create_realistic_noise_model(self) -> NoiseModel:
+        """Create comprehensive noise model for realistic simulation"""
+        noise_model = NoiseModel()
+
+        # Amplitude damping (T1 process)
+        t1_error = amplitude_damping_error(self.gate_time_1q / self.T1)
+        t1_error_2q = amplitude_damping_error(self.gate_time_2q / self.T1)
+
+        # Phase damping (T2 process)
+        t2_error = phase_damping_error(self.gate_time_1q / self.T2)
+        t2_error_2q = phase_damping_error(self.gate_time_2q / self.T2)
+
+        # Depolarizing errors (gate imperfections)
+        depol_1q = depolarizing_error(self.gate_error_1q, 1)
+        depol_2q = depolarizing_error(self.gate_error_2q, 2)
+
+        # Add errors to gates
+        noise_model.add_all_qubit_quantum_error(t1_error.compose(t2_error).compose(depol_1q),
+                                              ['h', 'x', 'y', 'z', 'rx', 'ry', 'rz', 'u'])
+        noise_model.add_all_qubit_quantum_error(t1_error_2q.compose(t2_error_2q).compose(depol_2q),
+                                              ['cx', 'cz', 'swap'])
+
+        # Measurement error
+        readout_error_model = [[1-self.readout_error, self.readout_error],
+                              [self.readout_error, 1-self.readout_error]]
+        noise_model.add_readout_error(readout_error_model, [0, 1, 2])
+
+        return noise_model
