+++ quantum_localization_demo.py
@@ -24,8 +24,24 @@ from typing import Dict, List, Tuple, Optional
 from dataclasses import dataclass
 from enum import Enum
 import json
+import argparse
 
-from src.quantum_localization_enhanced import QuantumLocalizationSystem
+from src.quantum_localization_enhanced import QuantumLocalizationSystem, run_technical_simulation
+from darpa_proposal_generator.theory import QuantumLocalizationTheory, RealisticQuantumSystem
+
+class NumpyJSONEncoder(json.JSONEncoder):
+    """
+    Custom JSON encoder for numpy data types.
+    This encoder handles numpy integers, floats, and arrays.
+    """
+    def default(self, obj):
+        if isinstance(obj, (np.integer, np.int_)):
+            return int(obj)
+        elif isinstance(obj, (np.floating, np.float_)):
+            return float(obj)
+        elif isinstance(obj, np.ndarray):
+            return obj.tolist()
+        return super(NumpyJSONEncoder, self).default(obj)
 
 # Enhanced logging with proper classification handling
 logging.basicConfig(
@@ -36,993 +52,11 @@ logging.basicConfig(
         logging.FileHandler('darpa_eris_enhanced_analysis.log')
     ]
 )
+from darpa_proposal_generator.analysis import EnhancedDARPAAnalysis, calculate_darpa_eris_score
+
+from darpa_proposal_generator.roadmap import ExperimentalValidationRoadmap
+
 logger = logging.getLogger(__name__)
-
-class QuantumLocalizationTheory:
-    """
-    Rigorous theoretical foundation for vibrational state localization
-    
-    Based on the mathematical framework where spatial coordinates are encoded
-    in the quantum harmonic oscillator basis states, allowing position
-    information to be manipulated through vibrational quantum numbers.
-    """
-    
-    def __init__(self, omega: float = 1.0, hbar: float = 1.0, mass: float = 1.0):
-        """
-        Initialize theoretical framework parameters
-        
-        Args:
-            omega: Harmonic oscillator frequency
-            hbar: Reduced Planck constant (natural units)
-            mass: Effective mass parameter
-        """
-        self.omega = omega
-        self.hbar = hbar
-        self.mass = mass
-        self.x0 = np.sqrt(hbar / (mass * omega))  # Characteristic length scale
-        
-    def harmonic_oscillator_wavefunction(self, n: int, x: np.ndarray) -> np.ndarray:
-        """
-        Generate quantum harmonic oscillator wavefunction for state |n⟩
-        
-        ψₙ(x) = (mω/πℏ)^(1/4) * (1/√(2ⁿn!)) * Hₙ(x/x₀) * exp(-x²/(2x₀²))
-        """
-        # Normalization constant
-        norm = (self.mass * self.omega / (np.pi * self.hbar))**(1/4)
-        norm *= 1.0 / np.sqrt(2**n * np.math.factorial(n))
-        
-        # Dimensionless coordinate
-        xi = x / self.x0
-        
-        # Hermite polynomial
-        hermite_poly = hermite(n)
-        Hn = hermite_poly(xi)
-        
-        # Gaussian envelope
-        gaussian = np.exp(-xi**2 / 2)
-        
-        return norm * Hn * gaussian
-    
-    def vibrational_coordinate_encoding(self, position: float, max_n: int = 10) -> np.ndarray:
-        """
-        Encode spatial position as superposition of vibrational states
-        
-        |ψ(x₀)⟩ = Σₙ cₙ|n⟩ where cₙ are chosen to localize at position x₀
-        
-        This is the core theoretical innovation: position becomes a quantum
-        observable through the vibrational quantum number basis.
-        """
-        # Calculate coefficients to maximize localization at target position
-        coefficients = np.zeros(max_n + 1, dtype=complex)
-        
-        # Use coherent state approach: |α⟩ = e^(-|α|²/2) Σₙ (αⁿ/√n!)|n⟩
-        alpha = position / self.x0  # Dimensionless displacement parameter
-        
-        for n in range(max_n + 1):
-            coefficients[n] = (alpha**n / np.sqrt(np.math.factorial(n))) * np.exp(-abs(alpha)**2 / 2)
-        
-        # Normalize
-        norm = np.sqrt(np.sum(np.abs(coefficients)**2))
-        coefficients /= norm
-        
-        return coefficients
-    
-    def theoretical_position_uncertainty(self, coefficients: np.ndarray) -> float:
-        """
-        Calculate theoretical minimum position uncertainty for given state
-        
-        Δx = x₀ * √⟨n̂⟩ for harmonic oscillator states
-        """
-        n_values = np.arange(len(coefficients))
-        mean_n = np.sum(np.abs(coefficients)**2 * n_values)
-        return self.x0 * np.sqrt(mean_n + 0.5)  # Include zero-point motion
-
-class RealisticQuantumSystem:
-    """
-    Model realistic quantum system with proper decoherence and error sources
-    """
-    
-    def __init__(self, platform_type: str = "superconducting"):
-        """
-        Initialize with realistic parameters for different quantum platforms
-        """
-        self.platform_type = platform_type
-        self.setup_platform_parameters()
-        
-    def setup_platform_parameters(self):
-        """Set realistic parameters based on current quantum technology"""
-        
-        if self.platform_type == "superconducting":
-            # IBM/Google superconducting qubit parameters (2024 state-of-art)
-            self.T1 = 100e-6  # Energy relaxation time (100 μs)
-            self.T2 = 50e-6   # Dephasing time (50 μs)
-            self.gate_time_1q = 30e-9  # Single-qubit gate time (30 ns)
-            self.gate_time_2q = 200e-9  # Two-qubit gate time (200 ns)
-            self.gate_error_1q = 1e-4   # Single-qubit gate error
-            self.gate_error_2q = 5e-3   # Two-qubit gate error
-            self.readout_error = 2e-2   # Measurement error
-            self.operating_temp = 0.015  # 15 mK
-            
-        elif self.platform_type == "trapped_ion":
-            # IonQ/Honeywell trapped ion parameters
-            self.T1 = 10.0    # Very long coherence (10 s)
-            self.T2 = 1.0     # Dephasing time (1 s)
-            self.gate_time_1q = 10e-6   # Single-qubit gate time (10 μs)
-            self.gate_time_2q = 100e-6  # Two-qubit gate time (100 μs)
-            self.gate_error_1q = 1e-5   # Excellent single-qubit fidelity
-            self.gate_error_2q = 1e-3   # Good two-qubit fidelity
-            self.readout_error = 1e-3   # Excellent readout
-            self.operating_temp = 1e-6  # μK effective temperature
-            
-        elif self.platform_type == "photonic":
-            # Xanadu/PsiQuantum photonic parameters
-            self.T1 = np.inf  # No energy relaxation for photons
-            self.T2 = 1e-3    # Limited by detection efficiency
-            self.gate_time_1q = 1e-12   # Speed of light limited (1 ps)
-            self.gate_time_2q = 1e-9    # Limited by nonlinear optics (1 ns)
-            self.gate_error_1q = 1e-3   # Limited by imperfect components
-            self.gate_error_2q = 1e-1   # Challenging two-qubit gates
-            self.readout_error = 1e-1   # Detector efficiency ~90%
-            self.operating_temp = 300   # Room temperature operation
-            
-        logger.info(f"Initialized {self.platform_type} quantum system with T1={self.T1:.2e}s, T2={self.T2:.2e}s")
-    
-    def create_realistic_noise_model(self) -> NoiseModel:
-        """Create comprehensive noise model for realistic simulation"""
-        noise_model = NoiseModel()
-        
-        # Amplitude damping (T1 process)
-        t1_error = amplitude_damping_error(self.gate_time_1q / self.T1)
-        t1_error_2q = amplitude_damping_error(self.gate_time_2q / self.T1)
-        
-        # Phase damping (T2 process)
-        t2_error = phase_damping_error(self.gate_time_1q / self.T2)
-        t2_error_2q = phase_damping_error(self.gate_time_2q / self.T2)
-        
-        # Depolarizing errors (gate imperfections)
-        depol_1q = depolarizing_error(self.gate_error_1q, 1)
-        depol_2q = depolarizing_error(self.gate_error_2q, 2)
-        
-        # Add errors to gates
-        noise_model.add_all_qubit_quantum_error(t1_error.compose(t2_error).compose(depol_1q), 
-                                              ['h', 'x', 'y', 'z', 'rx', 'ry', 'rz', 'u'])
-        noise_model.add_all_qubit_quantum_error(t1_error_2q.compose(t2_error_2q).compose(depol_2q),
-                                              ['cx', 'cz', 'swap'])
-        
-        # Measurement error
-        readout_error_model = [[1-self.readout_error, self.readout_error],
-                              [self.readout_error, 1-self.readout_error]]
-        noise_model.add_readout_error(readout_error_model, [0, 1, 2])
-        
-        return noise_model
-
-class EnhancedDARPAAnalysis:
-    """
-    Comprehensive DARPA ERIS analysis assessment
-    """
-    
-    def __init__(self, quantum_platform: str = "superconducting"):
-        self.quantum_system = RealisticQuantumSystem(quantum_platform)
-        self.theory = QuantumLocalizationTheory()
-        self.platform = quantum_platform
-        
-        # Conservative TRL assessment
-        self.current_trl = 2  # Technology concept formulated
-        self.target_trl_18m = 3  # Proof of concept (realistic)
-        self.target_trl_36m = 4  # Lab validation (conservative)
-        
-        logger.info(f"Enhanced DARPA analysis initialized for {quantum_platform} platform")
-    
-    def rigorous_problem_analysis(self) -> Dict:
-        """Enhanced problem analysis with detailed market research"""
-        
-        # Updated market data with sources
-        gps_vulnerability_data = {
-            "jamming_incidents_2023": 14850,  # Based on GNSS interference reports
-            "spoofing_incidents_2023": 2340,   # Maritime and aviation spoofing
-            "cost_per_incident_avg": 167000,   # DoD estimate per GPS disruption
-            "total_annual_cost": 2.5e9,        # Conservative DoD estimate
-            "affected_platforms": {
-                "submarines": 68,              # US Navy submarine fleet
-                "surface_ships": 290,         # Active fleet size
-                "aircraft": 13000,            # Total military aircraft
-                "ground_vehicles": 250000,    # Ground vehicle fleet
-                "precision_munitions": 50000  # Annual PGM production
-            }
-        }
-        
-        # Current technology limitations (verified data)
-        current_sota_limitations = {
-            "gps": {
-                "accuracy_degraded": 30.0,    # 30m in contested environments
-                "jamming_vulnerable": True,
-                "spoofing_vulnerable": True,
-                "availability_contested": 0.15  # 15% availability under attack
-            },
-            "ins": {
-                "drift_rate_mems": 1.0,       # 1 m/hr for MEMS INS
-                "drift_rate_ring_laser": 0.1, # 0.1 m/hr for ring laser gyro
-                "cost_mems": 1000,            # $1k for MEMS
-                "cost_ring_laser": 100000,    # $100k for ring laser
-                "no_external_ref": True
-            },
-            "celestial": {
-                "accuracy_clear": 50.0,       # 50m under ideal conditions
-                "accuracy_cloudy": 500.0,     # 500m with cloud cover
-                "availability_weather": 0.6,  # 60% weather-dependent availability
-                "update_rate": 0.1            # 0.1 Hz maximum update rate
-            }
-        }
-        
-        problem_analysis = {
-            "primary_challenge": "GPS denial in contested electromagnetic environments",
-            "quantified_impact": gps_vulnerability_data,
-            "current_limitations": current_sota_limitations,
-            "military_scenarios": {
-                "submarine_arctic": {
-                    "gps_availability": 0.0,   # No GPS under ice
-                    "ins_accuracy_required": 10.0,  # 10m for safe navigation
-                    "mission_duration_hours": 720    # 30-day patrol
-                },
-                "aircraft_ew_environment": {
-                    "gps_availability": 0.2,   # 20% availability under jamming
-                    "accuracy_required": 1.0,  # 1m for precision strikes
-                    "update_rate_required": 10.0  # 10 Hz for flight control
-                },
-                "ground_urban_canyon": {
-                    "gps_accuracy_degraded": 50.0,  # 50m multipath error
-                    "accuracy_required": 3.0,        # 3m for dismounted ops
-                    "battery_life_constraint": 72    # 72-hour operation
-                }
-            },
-            "technology_gap_analysis": {
-                "quantum_positioning_systems": "TRL 2-3 globally",
-                "commercial_quantum_sensors": "Limited to laboratory demonstrations",
-                "military_quantum_programs": "Early research phase",
-                "integration_challenges": "No quantum-classical interfaces deployed"
-            }
-        }
-        
-        return problem_analysis
-    
-    def realistic_quantum_advantage_assessment(self, num_trials: int = 1000) -> Dict:
-        """Conservative quantum advantage analysis with realistic error models"""
-        
-        logger.info("Performing realistic quantum advantage assessment...")
-        
-        # Classical system performance (based on real data)
-        classical_performance = {
-            "ins_drift_rate": 1.0,          # m/hr for tactical grade MEMS
-            "gps_accuracy_clear": 3.0,      # 3m in clear conditions
-            "gps_accuracy_contested": 30.0, # 30m under jamming
-            "gps_availability_contested": 0.15,  # 15% under attack
-            "celestial_accuracy": 50.0,     # 50m typical
-            "celestial_availability": 0.6   # 60% weather dependent
-        }
-        
-        # Quantum system theoretical performance
-        quantum_performance = {
-            "theoretical_accuracy": 0.1,    # Sub-wavelength theoretical limit
-            "decoherence_limited_accuracy": 1.0,  # Realistic with decoherence
-            "jamming_immunity": True,        # Fundamental quantum property
-            "spoofing_immunity": True,       # Quantum authentication
-            "update_rate": 1.0              # 1 Hz realistic for early systems
-        }
-        
-        # Monte Carlo simulation with realistic noise
-        noise_model = self.quantum_system.create_realistic_noise_model()
-        
-        quantum_errors = []
-        classical_errors = []
-        
-        for trial in range(num_trials):
-            # Classical system error (INS + occasional GPS)
-            mission_hours = np.random.uniform(1, 24)  # 1-24 hour missions
-            
-            if np.random.random() < classical_performance["gps_availability_contested"]:
-                # GPS available - use GPS accuracy
-                classical_error = np.random.normal(classical_performance["gps_accuracy_contested"], 5.0)
-            else:
-                # GPS denied - INS drift
-                classical_error = classical_performance["ins_drift_rate"] * mission_hours
-                classical_error += np.random.normal(0, classical_error * 0.1)  # 10% additional uncertainty
-            
-            classical_errors.append(abs(classical_error))
-            
-            # Quantum system error (including realistic decoherence)
-            # Base quantum accuracy limited by decoherence
-            base_error = quantum_performance["decoherence_limited_accuracy"]
-            
-            # Add platform-specific noise
-            platform_noise = {
-                "superconducting": 0.5,  # Additional error from thermal fluctuations
-                "trapped_ion": 0.1,      # Excellent coherence
-                "photonic": 0.3          # Limited by detection efficiency
-            }.get(self.platform, 0.5)
-            
-            quantum_error = np.sqrt(base_error**2 + platform_noise**2)
-            quantum_error += np.random.normal(0, quantum_error * 0.05)  # 5% measurement uncertainty
-            
-            quantum_errors.append(abs(quantum_error))
-        
-        quantum_errors = np.array(quantum_errors)
-        classical_errors = np.array(classical_errors)
-        
-        # Calculate realistic quantum advantage
-        advantage_metrics = {
-            "classical_mean_error_m": np.mean(classical_errors),
-            "quantum_mean_error_m": np.mean(quantum_errors),
-            "accuracy_improvement_factor": np.mean(classical_errors) / np.mean(quantum_errors),
-            "classical_95_percentile": np.percentile(classical_errors, 95),
-            "quantum_95_percentile": np.percentile(quantum_errors, 95),
-            "reliability_improvement": (np.sum(classical_errors > 10) - np.sum(quantum_errors > 10)) / len(classical_errors),
-            "quantum_advantage_factor": np.mean(classical_errors) / np.mean(quantum_errors),
-            "jamming_immunity": True,
-            "spoofing_immunity": True,
-            "power_consumption_ratio": 2.0,  # Quantum system uses 2x power initially
-            "size_weight_ratio": 3.0,        # 3x larger initially
-            "cost_ratio": 10.0               # 10x more expensive initially
-        }
-        
-        # Theoretical advantages that are fundamental
-        fundamental_advantages = {
-            "information_theoretic_security": "Quantum key distribution principles",
-            "measurement_back_action": "Eavesdropping detection through quantum mechanics",
-            "entanglement_enhanced_sensing": "Heisenberg-limited phase estimation",
-            "distributed_quantum_sensing": "Clock synchronization via entanglement"
-        }
-        
-        advantage_metrics["fundamental_advantages"] = fundamental_advantages
-        
-        logger.info(f"Quantum advantage assessment complete. Accuracy improvement: {advantage_metrics['accuracy_improvement_factor']:.1f}x")
-        
-        return advantage_metrics
-    
-    def enhanced_team_assessment(self) -> Dict:
-        """Detailed team capability assessment with realistic experience levels"""
-        
-        team_assessment = {
-            "principal_investigator": {
-                "name": "Christopher Woodyard",
-                "credentials": "CEO and Founder, Vers3Dynamics",
-                "quantum_experience_years": 3,
-                "relevant_publications": 0,  # Honest assessment
-                "relevant_patents": 0,       # Honest assessment
-                "military_experience": "Defense contractor integration",
-                "leadership_projects": ["AI research lab establishment", "Independent quantum research"],
-                "security_clearance": "None (Secret clearable)",
-                "strength_areas": [
-                    "AI and machine learning systems",
-                    "Software architecture and development",
-                    "Independent research initiative",
-                    "Military application understanding"
-                ],
-                "development_areas": [
-                    "Quantum hardware experience",
-                    "Peer-reviewed publication record",
-                    "Large team management experience",
-                    "Government contracting experience"
-                ]
-            },
-            "ai_rain_lab_structure": {
-                "quantum_hardware_specialist": {
-                    "required": True,
-                    "experience_level": "10+ years",
-                    "key_skills": ["Superconducting qubits", "Cryogenic systems", "RF control"],
-                    "recruitment_strategy": "Partner with MIT Lincoln Lab or IBM Quantum"
-                },
-                "navigation_systems_engineer": {
-                    "required": True,
-                    "experience_level": "15+ years",
-                    "key_skills": ["INS systems", "Kalman filtering", "Military navigation"],
-                    "recruitment_strategy": "Retired Navy/Air Force navigation specialist"
-                },
-                "quantum_software_architect": {
-                    "required": True,
-                    "experience_level": "5+ years",
-                    "key_skills": ["Qiskit/Cirq", "Quantum algorithms", "Error correction"],
-                    "recruitment_strategy": "Quantum computing startup experience"
-                },
-                "systems_integration_engineer": {
-                    "required": True,
-                    "experience_level": "10+ years",
-                    "key_skills": ["SWaP optimization", "Military standards", "Environmental testing"],
-                    "recruitment_strategy": "Defense aerospace background"
-                }
-            },
-            "ai_advisory_board": {
-                "quantum_theory_advisor": {
-                    "profile": "University professor with quantum sensing expertise",
-                    "commitment": "10 hours/month consultation",
-                    "compensation": "Equity + consulting fees"
-                },
-                "military_advisor": {
-                    "profile": "Retired flag officer with navigation systems experience",
-                    "commitment": "5 hours/month strategic guidance",
-                    "compensation": "Advisory fee"
-                },
-                "industry_advisor": {
-                    "profile": "Senior engineer from Lockheed Martin or Raytheon",
-                    "commitment": "10 hours/month technical review",
-                    "compensation": "Consulting agreement"
-                }
-            },
-            "organizational_capabilities": {
-                "current_capabilities": [
-                    "Small team agility and innovation",
-                    "Software development expertise",
-                    "Theoretical quantum research",
-                    "Military application focus"
-                ],
-                "required_capabilities": [
-                    "Quantum hardware development",
-                    "Military-grade system engineering",
-                    "Government contracting processes",
-                    "Security protocol implementation"
-                ],
-                "capability_gap_mitigation": {
-                    "partnerships": "National labs (Lincoln Lab, NIST)",
-                    "subcontracting": "Established defense contractors",
-                    "hiring": "Experienced quantum hardware team",
-                    "training": "Security clearance processing"
-                }
-            },
-            "execution_plan": {
-                "phase_1_18_months": {
-                    "team_size": 8,
-                    "key_hires": ["Quantum hardware lead", "Navigation systems engineer"],
-                    "deliverables": [
-                        "Theoretical framework validation",
-                        "Laboratory quantum localization demonstration",
-                        "Classical-quantum interface prototype"
-                    ],
-                    "success_metrics": [
-                        "10cm localization accuracy in lab",
-                        "1 second coherence time achievement",
-                        "Quantum advantage demonstration"
-                    ]
-                },
-                "phase_2_24_months": {
-                    "team_size": 15,
-                    "key_partnerships": ["MIT Lincoln Lab", "Defense contractor"],
-                    "deliverables": [
-                        "Ruggedized prototype system",
-                        "Military environment testing",
-                        "Integration with existing navigation systems"
-                    ],
-                    "success_metrics": [
-                        "1m accuracy in relevant environment",
-                        "Military temperature range operation",
-                        "TRL 4 validation by independent party"
-                    ]
-                }
-            },
-            "risk_assessment": {
-                "technical_risks": {
-                    "decoherence_limitations": {
-                        "probability": "Medium",
-                        "impact": "High",
-                        "mitigation": "Conservative performance targets, multiple platform approaches"
-                    },
-                    "integration_complexity": {
-                        "probability": "High", 
-                        "impact": "Medium",
-                        "mitigation": "Incremental integration approach, standard interfaces"
-                    }
-                },
-                "team_risks": {
-                    "key_personnel_availability": {
-                        "probability": "Medium",
-                        "impact": "High",
-                        "mitigation": "Competitive compensation, equity participation, flexible work"
-                    },
-                    "security_clearance_delays": {
-                        "probability": "High",
-                        "impact": "Medium", 
-                        "mitigation": "Early clearance processing, cleared staff partnerships"
-                    }
-                },
-                "programmatic_risks": {
-                    "funding_continuity": {
-                        "probability": "Low",
-                        "impact": "High",
-                        "mitigation": "Multiple funding sources, commercial applications"
-                    },
-                    "technology_readiness_timeline": {
-                        "probability": "Medium",
-                        "impact": "Medium",
-                        "mitigation": "Conservative milestones, parallel development paths"
-                    }
-                }
-            }
-        }
-        
-        return team_assessment
-    
-    def realistic_market_analysis(self) -> Dict:
-        """Conservative market analysis with verified data sources"""
-        
-        # Market data based on actual defense spending and industry reports
-        market_analysis = {
-            "defense_market": {
-                "navigation_systems_market_2024": 8.2e9,  # $8.2B (Verified: Defense industry reports)
-                "quantum_defense_market_2024": 1.1e9,     # $1.1B (IBM/McKinsey quantum report)
-                "projected_quantum_defense_2030": 8.6e9,  # $8.6B (Conservative growth)
-                
-                "addressable_segments": {
-                    "submarine_navigation": {
-                        "market_size": 450e6,      # $450M annually
-                        "key_players": ["Northrop Grumman", "L3Harris", "Thales"],
-                        "procurement_timeline": "5-7 years",
-                        "barriers_to_entry": "High (qualification requirements)"
-                    },
-                    "aircraft_navigation": {
-                        "market_size": 2.1e9,      # $2.1B annually  
-                        "key_players": ["Honeywell", "Collins Aerospace", "Garmin"],
-                        "procurement_timeline": "3-5 years",
-                        "barriers_to_entry": "Very High (DO-178C certification)"
-                    },
-                    "ground_vehicle_navigation": {
-                        "market_size": 800e6,      # $800M annually
-                        "key_players": ["BAE Systems", "General Dynamics", "Oshkosh"],
-                        "procurement_timeline": "2-4 years", 
-                        "barriers_to_entry": "Medium (MIL-STD compliance)"
-                    }
-                },
-                
-                "realistic_market_penetration": {
-                    "year_5": 0.001,   # 0.1% market penetration by year 5
-                    "year_10": 0.02,   # 2% market penetration by year 10
-                    "year_15": 0.08,   # 8% market penetration by year 15 (mature)
-                },
-                
-                "revenue_projections": {
-                    "conservative_scenario": {
-                        "year_5_revenue": 3.2e6,    # $3.2M
-                        "year_10_revenue": 66e6,    # $66M  
-                        "year_15_revenue": 260e6    # $260M
-                    },
-                    "optimistic_scenario": {
-                        "year_5_revenue": 8.5e6,    # $8.5M
-                        "year_10_revenue": 180e6,   # $180M
-                        "year_15_revenue": 680e6    # $680M
-                    }
-                }
-            },
-            
-            "commercial_market": {
-                "autonomous_vehicles": {
-                    "market_size_2030": 45e9,      # $45B (McKinsey)
-                    "quantum_addressable": 2.3e9,  # $2.3B (backup navigation)
-                    "timeline_to_market": "8-12 years"
-                },
-                "precision_agriculture": {
-                    "market_size_2030": 12e9,      # $12B
-                    "quantum_addressable": 600e6,  # $600M
-                    "timeline_to_market": "5-8 years"
-                },
-                "financial_timing": {
-                    "market_size_2030": 850e6,     # $850M (Bloomberg)
-                    "quantum_addressable": 85e6,   # $85M (quantum clocks)
-                    "timeline_to_market": "3-5 years"
-                }
-            },
-            
-            "competitive_landscape": {
-                "direct_competitors": {
-                    "cambridge_quantum_computing": {
-                        "focus": "Quantum software and algorithms",
-                        "quantum_sensing": "Limited",
-                        "military_contracts": "Small scale",
-                        "differentiation": "Software focus vs our hardware integration"
-                    },
-                    "qnami": {
-                        "focus": "Quantum sensing with NV centers",
-                        "navigation_focus": "None",
-                        "military_contracts": "Research only",
-                        "differentiation": "Materials sensing vs navigation"
-                    },
-                    "mu_space": {
-                        "focus": "Quantum gravimeters",
-                        "navigation_focus": "Indirect",
-                        "military_contracts": "Small",
-                        "differentiation": "Gravity sensing vs position sensing"
-                    }
-                },
-                "indirect_competitors": {
-                    "honeywell_aerospace": "Dominant in INS, no quantum capability",
-                    "northrop_grumman": "Advanced INS, some quantum research",
-                    "bae_systems": "Naval navigation systems, no quantum"
-                }
-            },
-            
-            "go_to_market_strategy": {
-                "phase_1_government": {
-                    "target_customers": ["DARPA", "ONR", "AFRL", "ARL"],
-                    "contract_vehicles": ["SBIR Phase I/II", "OTA agreements", "BAA responses"],
-                    "timeline": "Months 1-36",
-                    "revenue_target": 5e6  # $5M in government contracts
-                },
-                "phase_2_defense_prime": {
-                    "target_partners": ["Lockheed Martin", "Raytheon", "Northrop Grumman"],
-                    "partnership_model": "Technology licensing and co-development",
-                    "timeline": "Months 24-60",
-                    "revenue_target": 25e6  # $25M in prime contractor partnerships
-                },
-                "phase_3_commercial": {
-                    "target_markets": ["Autonomous vehicles", "Precision agriculture", "Financial timing"],
-                    "business_model": "Direct sales and licensing",
-                    "timeline": "Months 48-120",
-                    "revenue_target": 100e6  # $100M commercial revenue
-                }
-            },
-            
-            "investment_requirements": {
-                "total_development_cost": 45e6,  # $45M over 5 years (conservative)
-                "breakdown": {
-                    "personnel_60_percent": 27e6,
-                    "equipment_25_percent": 11.25e6,
-                    "facilities_10_percent": 4.5e6,
-                    "other_5_percent": 2.25e6
-                },
-                "funding_sources": {
-                    "government_contracts": 25e6,  # 55% government funding
-                    "private_investment": 15e6,    # 33% private investment
-                    "company_investment": 5e6      # 12% company resources
-                },
-                "roi_projections": {
-                    "break_even_year": 7,
-                    "5_year_roi": 0.15,    # 15% ROI by year 5
-                    "10_year_roi": 2.8     # 280% ROI by year 10
-                }
-            }
-        }
-        
-        return market_analysis
-
-def calculate_darpa_eris_score(problem_analysis: Dict, 
-                             quantum_advantage: Dict,
-                             team_assessment: Dict, 
-                             market_analysis: Dict) -> float:
-    """
-    Calculate a plausible DARPA ERIS score based on analysis dictionaries.
-    The score is out of 100, with weightings based on the report generation.
-    """
-
-    scores = {}
-
-    # 1. Problem Definition Score (40 points)
-    # Score based on quantified impact. Let's use total annual cost.
-    # Max score if cost is > $2B.
-    problem_score = min(problem_analysis['quantified_impact']['total_annual_cost'] / 2e9, 1.0) * 40
-    scores['problem'] = problem_score
-
-    # 2. State-of-Art Advancement Score (40 points)
-    # Score based on quantum advantage factor. Max score if factor is > 10x.
-    advantage_score = min(quantum_advantage['accuracy_improvement_factor'] / 10.0, 1.0) * 40
-    scores['advantage'] = advantage_score
-
-    # 3. Team Capability Score (15 points)
-    # Subjective score. Let's check PI's strengths vs weaknesses and hiring plan.
-    pi_assessment = team_assessment['principal_investigator']
-    num_strengths = len(pi_assessment['strength_areas'])
-    num_weaknesses = len(pi_assessment['development_areas'])
-    team_score = (num_strengths / (num_strengths + num_weaknesses))
-    # Bonus for having a good hiring plan
-    if team_assessment['ai_rain_lab_structure']['quantum_hardware_specialist']['required']:
-        team_score += 0.2
-    team_score = min(team_score, 1.0) * 15
-    scores['team'] = team_score
-
-    # 4. Market Impact Score (5 points)
-    # Based on 10-year ROI. Max score if ROI > 2.5
-    roi = market_analysis['investment_requirements']['roi_projections']['10_year_roi']
-    market_score = min(roi / 2.5, 1.0) * 5
-    scores['market'] = market_score
-    
-    total_score = sum(scores.values())
-
-    return total_score
-
-class ExperimentalValidationRoadmap:
-    """
-    Comprehensive experimental validation plan with realistic milestones
-    """
-    
-    def __init__(self):
-        self.current_trl = 2
-        self.validation_phases = self.define_validation_phases()
-    
-    def define_validation_phases(self) -> Dict:
-        """Define realistic experimental validation roadmap"""
-        
-        phases = {
-            "phase_1_theoretical_validation": {
-                "duration_months": 6,
-                "trl_start": 2,
-                "trl_end": 3,
-                "budget": 1.5e6,
-                "key_experiments": [
-                    {
-                        "experiment": "Harmonic oscillator state preparation",
-                        "platform": "Superconducting qubits (IBM Quantum)",
-                        "success_metric": "99% state preparation fidelity",
-                        "timeline_months": 3,
-                        "risk_level": "Low"
-                    },
-                    {
-                        "experiment": "Vibrational state superposition",
-                        "platform": "Trapped ions (IonQ access)",
-                        "success_metric": "Coherent superposition of 5+ vibrational modes",
-                        "timeline_months": 4,
-                        "risk_level": "Medium"
-                    },
-                    {
-                        "experiment": "Position encoding demonstration",
-                        "platform": "Quantum simulator (classical)",
-                        "success_metric": "Theoretical 10cm localization accuracy",
-                        "timeline_months": 2,
-                        "risk_level": "Low"
-                    }
-                ],
-                "deliverables": [
-                    "Peer-reviewed publication on vibrational localization theory",
-                    "Quantum circuit implementations for position encoding",
-                    "Simulation framework validation"
-                ],
-                "success_criteria": "All key experiments achieve >80% of target metrics"
-            },
-            
-            "phase_2_laboratory_demonstration": {
-                "duration_months": 12,
-                "trl_start": 3,
-                "trl_end": 4,
-                "budget": 8e6,
-                "key_experiments": [
-                    {
-                        "experiment": "Quantum localization in controlled environment",
-                        "platform": "Custom superconducting qubit system",
-                        "success_metric": "1m localization accuracy in lab",
-                        "timeline_months": 8,
-                        "risk_level": "High"
-                    },
-                    {
-                        "experiment": "Decoherence characterization",
-                        "platform": "Multiple quantum platforms",
-                        "success_metric": "100μs coherence time for localization",
-                        "timeline_months": 6,
-                        "risk_level": "Medium"
-                    },
-                    {
-                        "experiment": "Classical-quantum interface",
-                        "platform": "Hybrid classical-quantum system",
-                        "success_metric": "Real-time position updates at 1Hz",
-                        "timeline_months": 10,
-                        "risk_level": "Medium"
-                    },
-                    {
-                        "experiment": "Noise resilience testing",
-                        "platform": "Laboratory with controlled noise",
-                        "success_metric": "Maintains accuracy under 30dB interference",
-                        "timeline_months": 9,
-                        "risk_level": "Medium"
-                    }
-                ],
-                "deliverables": [
-                    "Laboratory prototype quantum localization system",
-                    "Performance characterization report",
-                    "Integration interface specifications"
-                ],
-                "success_criteria": "Demonstrate quantum advantage in controlled laboratory environment"
-            },
-            
-            "phase_3_relevant_environment": {
-                "duration_months": 18,
-                "trl_start": 4,
-                "trl_end": 5,
-                "budget": 15e6,
-                "key_experiments": [
-                    {
-                        "experiment": "Environmental robustness testing",
-                        "platform": "Ruggedized prototype",
-                        "success_metric": "Operation in -40°C to +70°C range",
-                        "timeline_months": 12,
-                        "risk_level": "High"
-                    },
-                    {
-                        "experiment": "Mobile platform integration",
-                        "platform": "Vehicle-mounted system",
-                        "success_metric": "3m accuracy during vehicle motion",
-                        "timeline_months": 15,
-                        "risk_level": "High"
-                    },
-                    {
-                        "experiment": "GPS-denied environment testing",
-                        "platform": "Underground/shielded facility",
-                        "success_metric": "Maintain localization without GPS",
-                        "timeline_months": 14,
-                        "risk_level": "Medium"
-                    },
-                    {
-                        "experiment": "Electromagnetic interference testing",
-                        "platform": "Anechoic chamber with RF sources",
-                        "success_metric": "Immune to 60dB jamming signals",
-                        "timeline_months": 10,
-                        "risk_level": "Medium"
-                    }
-                ],
-                "deliverables": [
-                    "Ruggedized quantum localization prototype",
-                    "Military environment test report",
-                    "Integration guidelines for defense platforms"
-                ],
-                "success_criteria": "Demonstrate reliable operation in relevant military environments"
-            },
-            
-            "phase_4_operational_environment": {
-                "duration_months": 24,
-                "trl_start": 5,
-                "trl_end": 6,
-                "budget": 20e6,
-                "key_experiments": [
-                    {
-                        "experiment": "Submarine navigation trial",
-                        "platform": "Navy submarine (if approved)",
-                        "success_metric": "10m accuracy over 24-hour dive",
-                        "timeline_months": 20,
-                        "risk_level": "Very High"
-                    },
-                    {
-                        "experiment": "Aircraft navigation trial",
-                        "platform": "Military aircraft testbed",
-                        "success_metric": "5m accuracy during flight operations",
-                        "timeline_months": 18,
-                        "risk_level": "Very High"
-                    },
-                    {
-                        "experiment": "Ground vehicle navigation trial",
-                        "platform": "Army ground vehicle",
-                        "success_metric": "3m accuracy in urban environment",
-                        "timeline_months": 15,
-                        "risk_level": "High"
-                    }
-                ],
-                "deliverables": [
-                    "Operational prototype systems",
-                    "Military field test results",
-                    "Technology transfer package"
-                ],
-                "success_criteria": "Successful operational demonstrations with military partners"
-            }
-        }
-        
-        return phases
-    
-    def generate_risk_mitigation_plan(self) -> Dict:
-        """Generate comprehensive risk mitigation strategies"""
-        
-        risk_mitigation = {
-            "technical_risks": {
-                "decoherence_limitations": {
-                    "risk_description": "Quantum coherence times insufficient for practical navigation",
-                    "probability": "Medium (40%)",
-                    "impact": "High (delays deployment by 12-24 months)",
-                    "mitigation_strategies": [
-                        "Parallel development on multiple quantum platforms",
-                        "Conservative performance targets with margin",
-                        "Hybrid classical-quantum algorithms",
-                        "Error correction protocol development"
-                    ],
-                    "contingency_plan": "Fall back to quantum-enhanced classical navigation",
-                    "early_warning_indicators": [
-                        "Coherence times <50μs in lab conditions",
-                        "Decoherence rate increases >10x in mobile platform"
-                    ]
-                },
-                "integration_complexity": {
-                    "risk_description": "Quantum system integration with classical navigation proves too complex",
-                    "probability": "High (60%)",
-                    "impact": "Medium (increases cost by 25-50%)",
-                    "mitigation_strategies": [
-                        "Modular system architecture",
-                        "Standard interface development",
-                        "Incremental integration approach",
-                        "Classical backup systems"
-                    ],
-                    "contingency_plan": "Standalone quantum navigation system",
-                    "early_warning_indicators": [
-                        "Interface development >6 months behind schedule",
-                        "Power consumption >5x classical systems"
-                    ]
-                },
-                "manufacturing_scalability": {
-                    "risk_description": "Quantum hardware cannot be manufactured at scale",
-                    "probability": "Medium (30%)",
-                    "impact": "High (limits market penetration)",
-                    "mitigation_strategies": [
-                        "Partnership with quantum hardware manufacturers",
-                        "Standardized component selection",
-                        "Volume manufacturing agreements",
-                        "Alternative fabrication approaches"
-                    ],
-                    "contingency_plan": "Limited production for niche applications",
-                    "early_warning_indicators": [
-                        "Unit costs >$1M per system",
-                        "Manufacturing yield <50%"
-                    ]
-                }
-            },
-            "programmatic_risks": {
-                "funding_continuity": {
-                    "risk_description": "Government funding reduced or cancelled",
-                    "probability": "Low (20%)",
-                    "impact": "Very High (program termination)",
-                    "mitigation_strategies": [
-                        "Multiple funding sources (DARPA, ONR, AFRL, ARL)",
-                        "Private investor engagement",
-                        "Commercial application development",
-                        "International partnership opportunities"
-                    ],
-                    "contingency_plan": "Transition to commercial-focused development",
-                    "early_warning_indicators": [
-                        "Budget cuts >25% in any fiscal year",
-                        "Program manager changes >2 times"
-                    ]
-                },
-                "team_retention": {
-                    "risk_description": "Key technical personnel leave for other opportunities",
-                    "probability": "Medium (35%)",
-                    "impact": "Medium (delays and knowledge loss)",
-                    "mitigation_strategies": [
-                        "Competitive compensation packages",
-                        "Equity participation for key personnel",
-                        "Professional development opportunities",
-                        "Flexible work arrangements"
-                    ],
-                    "contingency_plan": "Rapid hiring and knowledge transfer protocols",
-                    "early_warning_indicators": [
-                        "Turnover rate >20% annually",
-                        "Key personnel expressing dissatisfaction"
-                    ]
-                },
-                "competition_advancement": {
-                    "risk_description": "Competitors achieve similar capabilities first",
-                    "probability": "Medium (40%)",
-                    "impact": "High (reduced market opportunity)",
-                    "mitigation_strategies": [
-                        "Accelerated development timeline",
-                        "Patent protection strategy",
-                        "First-mover advantage in military applications",
-                        "Unique differentiating features"
-                    ],
-                    "contingency_plan": "Focus on superior performance or cost advantages",
-                    "early_warning_indicators": [
-                        "Competitor quantum navigation announcements",
-                        "Similar DARPA awards to other teams"
-                    ]
-                }
-            },
-            "market_risks": {
-                "military_adoption_rate": {
-                    "risk_description": "Military adoption slower than projected",
-                    "probability": "High (70%)",
-                    "impact": "Medium (reduced revenue timeline)",
-                    "mitigation_strategies": [
-                        "Conservative adoption projections",
-                        "Early military partner engagement",
-                        "Demonstration of clear operational advantages",
-                        "Integration with existing procurement programs"
-                    ],
-                    "contingency_plan": "Focus on civilian applications first",
-                    "early_warning_indicators": [
-                        "Military feedback indicates low interest",
-                        "Procurement timeline >2 years longer than expected"
-                    ]
-                }
-            }
-        }
-        
-        return risk_mitigation
 
 def run_quick_demo() -> bool:
     """
@@ -1061,6 +95,14 @@ def run_enhanced_darpa_analysis():
     """
     Execute comprehensive enhanced DARPA ERIS analysis
     """
+    logger.info("="*80)
+    logger.info("STEP 1: Running Core Technical Simulations & Visualizations")
+    logger.info("="*80)
+    technical_results = run_technical_simulation()
+    logger.info("="*80)
+    logger.info("STEP 2: Running High-Level DARPA Proposal Analysis")
+    logger.info("="*80)
+
     logger.info("="*80)
     logger.info("ENHANCED QUANTUM NAVIGATION SYSTEM ANALYSIS v2.0")
     logger.info("="*80)
@@ -1164,470 +206,51 @@ def run_enhanced_darpa_analysis():
         'darpa_score': best_score
     }
 
-def generate_enhanced_darpa_report(results: Dict, validation_roadmap: ExperimentalValidationRoadmap, 
-                                 risk_mitigation: Dict, darpa_score: float) -> str:
-    """DARPA report"""
-    
-    problem_analysis = results['problem_analysis']
-    quantum_advantage = results['quantum_advantage']
-    team_assessment = results['team_assessment']
-    market_analysis = results['market_analysis']
-    
-    report = f"""
-QUANTUM-ENHANCED NAVIGATION SYSTEM (QENS)
-=======================================================
-
-CLASSIFICATION: UNCLASSIFIED
-DISTRIBUTION: Approved for public release; distribution unlimited
-
-EXECUTIVE SUMMARY:
-The Enhanced Quantum-Enhanced Navigation System (QENS) addresses critical military navigation 
-challenges through a rigorously validated quantum localization approach. This submission 
-represents a significant advancement over our initial proposal, incorporating realistic 
-performance projections, comprehensive risk assessment, and conservative development timelines 
-based on current quantum technology capabilities.
-
-KEY IMPROVEMENTS IN THIS ENHANCED SUBMISSION:
-• Realistic quantum advantage projections ({quantum_advantage['accuracy_improvement_factor']:.1f}x vs classical)
-• Conservative TRL progression (2→4 in 36 months vs original 2→5 in 18 months)
-• Comprehensive experimental validation roadmap
-• Detailed risk mitigation strategies
-• Verified market data and competitive analysis
-
-1. ENHANCED PROBLEM DEFINITION & STATE OF THE ART
-===============================================
-
-QUANTIFIED MILITARY PROBLEM SCOPE:
-The GPS vulnerability crisis represents a {problem_analysis['quantified_impact']['total_annual_cost']/1e9:.1f} billion dollar 
-annual impact to DoD operations, with {problem_analysis['quantified_impact']['jamming_incidents_2023']:,} documented 
-jamming incidents and {problem_analysis['quantified_impact']['spoofing_incidents_2023']:,} spoofing incidents in 2023 alone.
-
-AFFECTED MILITARY PLATFORMS:
-• Submarines: {problem_analysis['quantified_impact']['affected_platforms']['submarines']} vessels requiring GPS-denied navigation
-• Surface Ships: {problem_analysis['quantified_impact']['affected_platforms']['surface_ships']} vessels vulnerable to GPS attacks  
-• Aircraft: {problem_analysis['quantified_impact']['affected_platforms']['aircraft']:,} military aircraft requiring precise navigation
-• Ground Vehicles: {problem_analysis['quantified_impact']['affected_platforms']['ground_vehicles']:,} platforms needing backup navigation
-• Precision Munitions: {problem_analysis['quantified_impact']['affected_platforms']['precision_munitions']:,} annual PGM production affected
-
-COMPREHENSIVE STATE-OF-THE-ART ANALYSIS:
-
-Current System Performance (Verified Data):
-┌─────────────────┬──────────────┬───────────────┬─────────────────┬─────────────┐
-│ System          │ Accuracy     │ Jamming Vuln │ Cost/Unit       │ Availability│
-├─────────────────┼──────────────┼───────────────┼─────────────────┼─────────────┤
-│ GPS (Clear)     │ 3m           │ YES           │ $100            │ 95%         │
-│ GPS (Contested) │ 30m          │ YES           │ $100            │ 15%         │
-│ INS (MEMS)      │ 1m/hr drift  │ NO            │ $1,000          │ 100%        │
-│ INS (Ring Laser)│ 0.1m/hr drift│ NO            │ $100,000        │ 100%        │
-│ Celestial       │ 50m          │ NO            │ $10,000         │ 60%         │
-│ QENS (Proposed) │ 1-3m         │ NO            │ $100,000        │ 95%+        │
-└─────────────────┴──────────────┴───────────────┴─────────────────┴─────────────┘
-
-TECHNOLOGY GAPS IDENTIFIED:
-• No deployed quantum navigation systems exist globally (highest TRL is 3)
-• All jam-resistant systems suffer from significant accuracy degradation over time
-• Current quantum research lacks focus on military deployment requirements
-• No demonstrated quantum-classical navigation system integration
-
-2. REALISTIC ADVANCEMENT OF STATE OF THE ART
-==========================================
-
-CONSERVATIVE QUANTUM ADVANTAGE ASSESSMENT:
-Based on comprehensive Monte Carlo simulation with realistic noise models:
-
-Performance Metrics:
-• Accuracy Improvement: {quantum_advantage['accuracy_improvement_factor']:.1f}x over best classical alternative
-• Reliability Improvement: {quantum_advantage['reliability_improvement']:.1%} fewer navigation failures
-• Jamming Immunity: Fundamental quantum mechanical protection
-• Spoofing Immunity: Quantum authentication prevents false signals
-
-Realistic Performance Projections:
-• Laboratory Accuracy: 0.1m (Heisenberg-limited theoretical)
-• Field Accuracy (with decoherence): 1-3m (practical implementation)
-• Update Rate: 1Hz (conservative for Phase I systems)
-• Operating Temperature: -40°C to +70°C (military specification)
-
-FUNDAMENTAL QUANTUM ADVANTAGES:
-{chr(10).join([f"• {k}: {v}" for k, v in quantum_advantage.get('fundamental_advantages', {}).items()])}
-
-TECHNICAL INNOVATION BREAKTHROUGH:
-First practical implementation of vibrational state localization where spatial coordinates 
-are encoded in quantum harmonic oscillator basis states, enabling:
-• Position manipulation through quantum phase control
-• Inherent anti-jamming through quantum mechanical principles  
-• Distributed quantum sensor network capability
-• Quantum error correction for navigation states
-
-3. ENHANCED TEAM CAPABILITY ASSESSMENT
-====================================
-
-PRINCIPAL INVESTIGATOR: {team_assessment['principal_investigator']['name']}
-Current Experience: {team_assessment['principal_investigator']['quantum_experience_years']} years quantum systems
-Publications: {team_assessment['principal_investigator']['relevant_publications']} peer-reviewed (honest assessment)
-Patents: {team_assessment['principal_investigator']['relevant_patents']} relevant patents
-Security Clearance: {team_assessment['principal_investigator']['security_clearance']}
-
-ACKNOWLEDGED CAPABILITY GAPS & MITIGATION:
-Current Limitations:
-• Limited quantum hardware experience (mitigate via national lab partnerships)
-• No prior government contracting experience (mitigate via prime contractor teaming)
-• Small team size for complex project (mitigate via strategic hiring plan)
-• Limited security clearance depth (mitigate via early clearance processing)
-
-TEAM DEVELOPMENT STRATEGY:
-Phase I Team Expansion (18 months):
-├── Quantum Hardware Lead: MIT Lincoln Lab partnership or IBM Quantum alumnus
-├── Navigation Systems Engineer: Retired Navy/Air Force navigation specialist  
-├── Systems Integration Engineer: Defense aerospace background (Lockheed/Raytheon)
-└── Quantum Software Architect: Qiskit/Cirq expert from quantum computing startup
-
-Advisory Board Engagement:
-├── Technical Advisor: University quantum sensing professor (10 hrs/month)
-├── Military Advisor: Retired flag officer with navigation expertise (5 hrs/month)
-└── Industry Advisor: Senior defense contractor engineer (10 hrs/month)
-
-RISK MITIGATION FOR TEAM LIMITATIONS:
-• Partnership agreements with MIT Lincoln Lab for quantum hardware expertise
-• Subcontracting relationships with established defense contractors
-• Competitive compensation packages including equity participation
-• Early security clearance processing for key personnel
-
-4. REALISTIC DEFENSE AND COMMERCIAL MARKET ANALYSIS
-=================================================
-
-DEFENSE MARKET OPPORTUNITY (Verified Data):
-Total Addressable Market: ${market_analysis['defense_market']['navigation_systems_market_2024']/1e9:.1f}B (2024 defense navigation market)
-Quantum Defense Segment: ${market_analysis['defense_market']['quantum_defense_market_2024']/1e9:.1f}B (emerging quantum defense applications)
-
-CONSERVATIVE REVENUE PROJECTIONS:
-Year 5 Revenue: ${market_analysis['defense_market']['revenue_projections']['conservative_scenario']['year_5_revenue']/1e6:.1f}M (0.1% market penetration)
-Year 10 Revenue: ${market_analysis['defense_market']['revenue_projections']['conservative_scenario']['year_10_revenue']/1e6:.0f}M (2% market penetration)
-Year 15 Revenue: ${market_analysis['defense_market']['revenue_projections']['conservative_scenario']['year_15_revenue']/1e6:.0f}M (8% mature market penetration)
-
-REALISTIC COMPETITIVE POSITIONING:
-Current Competitors:
-• Cambridge Quantum Computing: Software focus, no navigation hardware
-• Qnami: NV center sensing, no navigation applications
-• Mu Space: Quantum gravimeters, indirect navigation relevance
-
-Competitive Advantages:
-• First-mover advantage in quantum navigation
-• Military-specific design requirements
-• Quantum-native anti-jamming capabilities
-• No licensing dependencies on foreign technology
-
-MARKET ENTRY STRATEGY:
-Phase 1 (Years 1-3): Government contracts (DARPA, ONR, AFRL) - ${market_analysis['go_to_market_strategy']['phase_1_government']['revenue_target']/1e6:.0f}M target
-Phase 2 (Years 2-5): Defense prime partnerships (Lockheed, Raytheon) - ${market_analysis['go_to_market_strategy']['phase_2_defense_prime']['revenue_target']/1e6:.0f}M target  
-Phase 3 (Years 4-10): Commercial applications (autonomous vehicles) - ${market_analysis['go_to_market_strategy']['phase_3_commercial']['revenue_target']/1e6:.0f}M target
-
-INVESTMENT REQUIREMENTS & ROI:
-Total Development Investment: ${market_analysis['investment_requirements']['total_development_cost']/1e6:.0f}M over 5 years
-Government Funding: ${market_analysis['investment_requirements']['funding_sources']['government_contracts']/1e6:.0f}M (55% of total)
-Private Investment: ${market_analysis['investment_requirements']['funding_sources']['private_investment']/1e6:.0f}M (33% of total)
-Break-even Timeline: {market_analysis['investment_requirements']['roi_projections']['break_even_year']} years
-10-Year ROI: {market_analysis['investment_requirements']['roi_projections']['10_year_roi']*100:.0f}% return on investment
-
-5. COMPREHENSIVE EXPERIMENTAL VALIDATION ROADMAP
-===============================================
-
-REALISTIC TRL PROGRESSION:
-Current State: TRL 2 (Technology concept and application formulated)
-18-Month Target: TRL 3 (Analytical and experimental proof of concept)
-36-Month Target: TRL 4 (Component validation in laboratory environment)
-
-PHASE 1: THEORETICAL VALIDATION (6 months, ${validation_roadmap.validation_phases['phase_1_theoretical_validation']['budget']/1e6:.1f}M)
-Key Experiments:
-├── Harmonic oscillator state preparation (IBM Quantum access)
-│   Success Metric: 99% state preparation fidelity
-│   Timeline: 3 months | Risk: Low
-├── Vibrational state superposition (IonQ partnership)  
-│   Success Metric: Coherent superposition of 5+ modes
-│   Timeline: 4 months | Risk: Medium
-└── Position encoding demonstration (quantum simulation)
-    Success Metric: Theoretical 10cm localization accuracy
-    Timeline: 2 months | Risk: Low
-
-PHASE 2: LABORATORY DEMONSTRATION (12 months, ${validation_roadmap.validation_phases['phase_2_laboratory_demonstration']['budget']/1e6:.0f}M)
-Key Experiments:
-├── Quantum localization in controlled environment
-│   Success Metric: 1m localization accuracy in lab
-│   Timeline: 8 months | Risk: High
-├── Decoherence characterization across platforms
-│   Success Metric: 100μs coherence time for localization
-│   Timeline: 6 months | Risk: Medium  
-├── Classical-quantum interface development
-│   Success Metric: Real-time position updates at 1Hz
-│   Timeline: 10 months | Risk: Medium
-└── Noise resilience testing
-    Success Metric: Maintains accuracy under 30dB interference
-    Timeline: 9 months | Risk: Medium
-
-PHASE 3: RELEVANT ENVIRONMENT TESTING (18 months, ${validation_roadmap.validation_phases['phase_3_relevant_environment']['budget']/1e6:.0f}M)
-Key Experiments:
-├── Environmental robustness (-40°C to +70°C operation)
-├── Mobile platform integration (vehicle-mounted testing)
-├── GPS-denied environment validation (underground facilities)
-└── Electromagnetic interference immunity (anechoic chamber testing)
-
-PHASE 4: OPERATIONAL ENVIRONMENT (24 months, ${validation_roadmap.validation_phases['phase_4_operational_environment']['budget']/1e6:.0f}M)
-Military Platform Testing:
-├── Submarine navigation trial (Navy partnership)
-├── Aircraft navigation trial (Air Force testbed)
-└── Ground vehicle trial (Army test vehicle)
-
-6. COMPREHENSIVE RISK ASSESSMENT & MITIGATION
-============================================
-
-TECHNICAL RISKS:
-High-Impact Technical Risk: Decoherence Limitations
-├── Probability: Medium (40%)
-├── Impact: High (12-24 month delay)
-├── Mitigation: Multiple quantum platforms, conservative targets, hybrid algorithms
-├── Contingency: Quantum-enhanced classical navigation fallback
-└── Early Warning: Coherence times <50μs, >10x decoherence in mobile platform
-
-Medium-Impact Technical Risk: Integration Complexity  
-├── Probability: High (60%)
-├── Impact: Medium (25-50% cost increase)
-├── Mitigation: Modular architecture, standard interfaces, incremental approach
-├── Contingency: Standalone quantum navigation system
-└── Early Warning: Interface development >6 months behind, power >5x classical
-
-PROGRAMMATIC RISKS:
-Critical Programmatic Risk: Funding Continuity
-├── Probability: Low (20%)  
-├── Impact: Very High (program termination)
-├── Mitigation: Multiple funding sources, commercial applications, international partnerships
-├── Contingency: Commercial-focused development transition
-└── Early Warning: Budget cuts >25%, >2 program manager changes
-
-Team Risk: Key Personnel Retention
-├── Probability: Medium (35%)
-├── Impact: Medium (delays and knowledge loss)
-├── Mitigation: Competitive compensation, equity participation, professional development
-├── Contingency: Rapid hiring and knowledge transfer protocols
-└── Early Warning: Turnover >20% annually, key personnel expressing dissatisfaction
-
-MARKET RISKS:
-Adoption Risk: Military Adoption Rate
-├── Probability: High (70%)
-├── Impact: Medium (reduced revenue timeline)
-├── Mitigation: Conservative projections, early military engagement, clear advantages
-├── Contingency: Civilian applications focus
-└── Early Warning: Low military feedback interest, procurement delays >2 years
-
-7. TECHNOLOGY READINESS AND REALISTIC MILESTONES
-==============================================
-
-CONSERVATIVE TRL ASSESSMENT:
-Current TRL 2 justification:
-✓ Theoretical framework established (vibrational state localization)
-✓ Initial quantum circuit designs developed
-✓ Mathematical models for quantum advantage validated
-✗ No experimental demonstration of key concepts
-✗ No hardware component validation
-✗ No system integration demonstrated
-
-Target TRL 3 (18 months):
-└── Experimental proof of concept demonstrations
-    ├── Vibrational state preparation and manipulation
-    ├── Position encoding in quantum states
-    ├── Decoherence characterization
-    └── Basic quantum advantage validation
-
-Target TRL 4 (36 months):  
-└── Component validation in laboratory environment
-    ├── Integrated quantum localization prototype
-    ├── Classical-quantum interface validation
-    ├── Environmental stability testing
-    └── Performance characterization
-
-REALISTIC MILESTONE TIMELINE:
-Month 6:   Theoretical validation complete, initial experiments begun
-Month 12:  Proof of concept demonstrations, TRL 3 achieved
-Month 18:  Laboratory prototype integration started
-Month 24:  Component validation testing, environmental characterization  
-Month 30:  System integration complete, performance validation
-Month 36:  TRL 4 demonstration, relevant environment testing begun
-
-8. ENHANCED INNOVATION SUMMARY
-============================
-
-CORE TECHNICAL INNOVATIONS:
-1. Vibrational State Coordinate Encoding
-   └── First practical implementation of position encoding in quantum harmonic oscillator basis
-   
-2. Quantum Phase Space Localization  
-   └── Novel approach using quantum phase manipulation for spatial coordinate transformation
-   
-3. Anti-Jamming Quantum Protection
-   └── Fundamental quantum mechanical immunity to classical electromagnetic interference
-   
-4. Distributed Quantum Sensor Networks
-   └── Scalable architecture for multi-platform quantum navigation networks
-
-INTELLECTUAL PROPERTY STRATEGY:
-• Core localization algorithms: Patent application filed
-• Vibrational state encoding: Patent pending  
-• Quantum-classical interfaces: Trade secret protection
-• Military integration protocols: Government use rights
-
-COMPETITIVE DIFFERENTIATION:
-• 3-5x accuracy improvement over jam-resistant alternatives
-• Complete immunity to electronic warfare attacks
-• No external dependencies or licensing fees
-• Quantum-secure by fundamental physics principles
-
-9. FINANCIAL PROJECTIONS AND RESOURCE REQUIREMENTS
-================================================
-
-PHASE-BY-PHASE BUDGET BREAKDOWN:
-Phase I (18 months): $8.5M
-├── Personnel (60%): $5.1M  
-├── Equipment (25%): $2.1M
-├── Facilities (10%): $0.85M
-└── Travel/Other (5%): $0.45M
-
-Phase II (24 months): $15.0M  
-├── Prototype development: $8.0M
-├── Military testing: $4.0M
-├── Integration: $2.0M
-└── Documentation: $1.0M
-
-Phase III (36 months): $21.5M
-├── Operational prototypes: $12.0M
-├── Field testing: $6.0M
-├── Manufacturing preparation: $2.5M
-└── Technology transfer: $1.0M
-
-Total Program Investment: $45.0M over 6 years
-
-FUNDING STRATEGY:
-Government Contracts (55%): $25.0M
-├── DARPA ERIS: $8.5M (Phase I)
-├── ONR Follow-on: $7.0M (Phase II)  
-├── AFRL Partnership: $5.0M (Phase III)
-└── ARL Collaboration: $4.5M (Ongoing)
-
-Private Investment (33%): $15.0M
-├── Seed funding: $2.0M (completed)
-├── Series A: $6.0M (Year 2)
-└── Series B: $7.0M (Year 4)
-
-Company Resources (12%): $5.0M
-├── Facilities and overhead
-├── Founder/team equity value
-└── In-kind contributions
-
-10. CONCLUSION AND RECOMMENDATIONS
-================================
-
-ENHANCED ASSESSMENT SUMMARY:
-This enhanced QENS proposal represents a significant improvement over initial submissions, 
-incorporating realistic performance projections, comprehensive risk assessment, and 
-conservative development timelines based on current quantum technology capabilities.
-
-DARPA ERIS SCORING BREAKDOWN:
-Problem Definition & State of Art: {darpa_score * 0.4:.1f}/40 points
-├── Clear $2.5B annual military problem quantification
-├── Comprehensive analysis of 4+ current alternatives
-└── Detailed technology gap identification
-
-Advancing State of Art: {darpa_score * 0.4:.1f}/40 points  
-├── {quantum_advantage['accuracy_improvement_factor']:.1f}x realistic quantum advantage demonstration
-├── 4 fundamental quantum advantages identified
-└── Novel vibrational state localization approach
-
-Team Capability: {darpa_score * 0.15:.1f}/15 points
-├── Honest assessment of current limitations
-├── Comprehensive capability gap mitigation strategy  
-└── Strong advisory board and partnership plan
-
-Defense/Commercial Impact: {darpa_score * 0.05:.1f}/5 points
-├── ${market_analysis['defense_market']['revenue_projections']['conservative_scenario']['year_10_revenue']/1e6:.0f}M conservative 10-year revenue projection
-├── Clear go-to-market strategy with military focus
-└── Realistic competitive positioning
-
-TOTAL ENHANCED DARPA ERIS SCORE: {darpa_score:.1f}/100
-
-RECOMMENDATION FOR DARPA AWARD:
-Based on this comprehensive enhanced analysis, QENS represents a viable quantum navigation 
-solution with realistic technical goals, conservative development timelines, and significant 
-potential military impact. The enhanced approach addresses previous concerns about 
-over-optimistic projections while maintaining the fundamental quantum advantages that 
-make this technology revolutionary for GPS-denied navigation.
-
-IMMEDIATE NEXT STEPS:
-1. DARPA ERIS Phase I award for 18-month proof of concept ($8.5M)
-2. National laboratory partnership agreements (MIT Lincoln Lab, NIST)
-3. Defense contractor teaming agreements (Lockheed Martin, Raytheon)
-4. Key personnel recruitment and security clearance processing
-5. Quantum hardware platform selection and procurement
-
-EXPECTED PROGRAM OUTCOMES:
-• TRL 4 quantum navigation system demonstration by Month 36
-• 1-3m navigation accuracy in GPS-denied environments
-• Fundamental quantum immunity to jamming and spoofing
-• Technology transfer package for defense contractor integration
-• Foundation for next-generation military navigation capabilities
-
-LONG-TERM STRATEGIC IMPACT:
-• U.S. quantum supremacy in navigation domain
-• $500M+ annual cost avoidance from GPS vulnerability reduction  
-• Enhanced military operational capability in contested environments
-• Commercial quantum navigation market leadership
-• International quantum technology export opportunities
-
-CLASSIFICATION: UNCLASSIFIED
-SUBMITTED BY: Christopher Woodyard, Principal Investigator
-ORGANIZATION: Vers3Dynamics R.A.I.N. Lab  
-ENHANCED SUBMISSION DATE: {time.strftime('%Y-%m-%d')}
-CONTACT: ciao_chris@proton.me
-
-=======================================================
-END OF ENHANCED DARPA ERIS SUBMISSION
-=======================================================
-"""
-
-    return report
+from darpa_proposal_generator.report import generate_enhanced_darpa_report
 
 
 # Main execution for enhanced analysis
 if __name__ == "__main__":
+    parser = argparse.ArgumentParser(description="Quantum-Enhanced Navigation System (QENS) Demo")
+    parser.add_argument(
+        '--full-analysis',
+        action='store_true',
+        help='Run the full, enhanced DARPA analysis.'
+    )
+    args = parser.parse_args()
+
     print("DARPA ERIS: Quantum-Enhanced Navigation System")
     print("=" * 70)
-    
-    # Run the quick demo by default
-    run_quick_demo()
 
-    # To run the full enhanced analysis, uncomment the following lines:
-    # try:
-    #     # Run enhanced comprehensive analysis
-    #     results = run_enhanced_darpa_analysis()
-        
-    #     print(f"\n✓ Enhanced DARPA ERIS analysis complete")
-    #     print(f"Final Enhanced Score: {results['darpa_score']:.1f}/100")
-        
-    #     # Save enhanced results
-    #     with open('enhanced_darpa_eris_submission.txt', 'w') as f:
-    #         f.write(results['enhanced_report'])
-    #     print("📄 Enhanced DARPA ERIS submission saved to 'enhanced_darpa_eris_submission.txt'")
-        
-    #     # Save detailed analysis data
-    #     with open('enhanced_analysis_data.json', 'w') as f:
-    #         # Convert non-serializable objects to serializable format
-    #         serializable_results = {
-    #             'darpa_score': results['darpa_score'],
-    #             'quantum_advantage': results['best_results']['quantum_advantage'],
-    #             'market_analysis': results['best_results']['market_analysis'],
-    #             'validation_phases': results['validation_roadmap'].validation_phases,
-    #             'risk_mitigation': results['risk_mitigation']
-    #         }
-    #         json.dump(serializable_results, f, indent=2, default=str)
-    #     print("📊 Detailed analysis data saved to 'enhanced_analysis_data.json'")
-        
-    # except Exception as e:
-    #     logger.error(f"Enhanced DARPA ERIS analysis failed: {str(e)}")
-    #     print(f"\n❌ Enhanced analysis failed: {str(e)}")
-    #     print("Check logs for detailed error information")
+    if args.full_analysis:
+        try:
+            # Run enhanced comprehensive analysis
+            results = run_enhanced_darpa_analysis()
+
+            print(f"\n✓ Enhanced DARPA ERIS analysis complete")
+            print(f"Final Enhanced Score: {results['darpa_score']:.1f}/100")
+
+            # Save enhanced results
+            with open('enhanced_darpa_eris_submission.txt', 'w') as f:
+                f.write(results['enhanced_report'])
+            print("📄 Enhanced DARPA ERIS submission saved to 'enhanced_darpa_eris_submission.txt'")
+
+            # Save detailed analysis data
+            with open('enhanced_analysis_data.json', 'w') as f:
+                serializable_results = {
+                    'darpa_score': results['darpa_score'],
+                    'quantum_advantage': results['best_results']['quantum_advantage'],
+                    'market_analysis': results['best_results']['market_analysis'],
+                    'validation_phases': results['validation_roadmap'].validation_phases,
+                    'risk_mitigation': results['risk_mitigation']
+                }
+                json.dump(serializable_results, f, indent=2, cls=NumpyJSONEncoder)
+            print("📊 Detailed analysis data saved to 'enhanced_analysis_data.json'")
+
+        except Exception as e:
+            logger.error(f"Enhanced DARPA ERIS analysis failed: {str(e)}")
+            print(f"\n❌ Enhanced analysis failed: {str(e)}")
+            print("Check logs for detailed error information")
+    else:
+        # Run the quick demo by default
+        run_quick_demo()
